rag-app/requirements.txt
faiss-cpu==1.8.0.post1; python_version < "3.13"
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
jinja2>=3.1.4
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.2
pydantic>=2.8.2
numpy>=1.26.4
pandas>=2.2.2
tiktoken>=0.7.0
typer>=0.12.3
rich>=13.7.1
sqlite-utils>=3.36
python-dotenv>=1.0.1
